So long and thanks for all the fish
So long and thanks for all the fish
So sad that it should come to this
We tried to warn you all but oh dear

You may not share our intellect
Which might explain your disrespect
For all the natural wonders that
grow around you

So long, so long and thanks
for all the fish

The world's about to be destroyed
There's no point getting all annoyed
Lie back and let the planet dissolve(around you)

Despite those nets of tuna fleets
We thought that most of you were sweet
Especially tiny tots and your
pregnant women

So long, so long, so long, so long, so long
So long, so long, so long, so long, so long

So long, so long and thanks
for all the fish

If I had just one last wish
I would like a tasty fish
If we could just change one thing
We would all have learned to sing.

Come one and all
Man and Mammal
Side by Side in life's great gene pool

(oooohhh oooohhh oooaahhhhh- ah ahh)

So long, so long, so long, so long, so long
So long, so long, so long, so long, so long

So long, so long and, !Thanks!
for all the fish!
(old change to make it easier to demo, and final push!? so long TED!)